+++ ohos_nweb/src/cef_delegate/nweb_delegate.cc
@@ -2430,7 +2430,7 @@ void NWebDelegate::RegisterAccessibilityEventListener(
 }
 
 void NWebDelegate::RegisterAccessibilityIdGenerator(
-    std::function<int32_t()> accessibilityIdGenerator) const {
+    std::function<int64_t()> accessibilityIdGenerator) const {
   content::BrowserAccessibilityManagerOHOS::RegisterAccessibilityIdGenerator(
       accessibilityIdGenerator);
 };
@@ -2444,7 +2444,7 @@ void NWebDelegate::SetAccessibilityState(cef_state_t accessibilityState) {
   GetBrowser()->GetHost()->SetAccessibilityState(accessibilityState);
 }
 
-void NWebDelegate::ExecuteAction(int32_t accessibilityId,
+void NWebDelegate::ExecuteAction(int64_t accessibilityId,
                                  uint32_t action) const {
   auto* accessibilityManager = GetAccessibilityManager();
   if (accessibilityManager == nullptr) {
@@ -2505,7 +2505,7 @@ NWebDelegate::GetAccessibilityManager() const {
 }
 
 bool NWebDelegate::GetFocusedAccessibilityNodeInfo(
-    int32_t accessibilityId,
+    int64_t accessibilityId,
     bool isAccessibilityFocus,
     OHOS::NWeb::NWebAccessibilityNodeInfo& nodeInfo) const {
   auto* accessibilityManager = GetAccessibilityManager();
@@ -2545,7 +2545,7 @@ bool NWebDelegate::GetFocusedAccessibilityNodeInfo(
 }
 
 bool NWebDelegate::GetAccessibilityNodeInfoById(
-    int32_t accessibilityId,
+    int64_t accessibilityId,
     OHOS::NWeb::NWebAccessibilityNodeInfo& nodeInfo) const {
   auto* accessibilityManager = GetAccessibilityManager();
   if (accessibilityManager == nullptr) {
@@ -2569,7 +2569,7 @@ bool NWebDelegate::GetAccessibilityNodeInfoById(
 }
 
 bool NWebDelegate::GetAccessibilityNodeInfoByFocusMove(
-    int32_t accessibilityId,
+    int64_t accessibilityId,
     int32_t direction,
     OHOS::NWeb::NWebAccessibilityNodeInfo& nodeInfo) const {
   auto* accessibilityManager = GetAccessibilityManager();
